LayoutTests/inspector/syntax-highlight-css.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="syntax-highlight.js"></script>
<script>

function test()
{
    function dumpSyntaxHighlightCSS(str)
    {
        InspectorTest.dumpSyntaxHighlight(str, "text/css");
    }

    dumpSyntaxHighlightCSS("a[href='/']");
    dumpSyntaxHighlightCSS("#content > a:hover");
    dumpSyntaxHighlightCSS("@import url(style.css);");
    dumpSyntaxHighlightCSS("@import url(\"style.css\") projection, tv;");
    dumpSyntaxHighlightCSS("@import \"/css/fireball_unicode.css\"; html {}");
    dumpSyntaxHighlightCSS("@media screen { body { color: red; } }");
    dumpSyntaxHighlightCSS("@font-face { font-family: \"MyHelvetica\"; }");
    dumpSyntaxHighlightCSS("p { color: color; red: red; color: #000; color: #FFF; color: #123AbC; color: #faebfe; color:papayawhip; }");
    dumpSyntaxHighlightCSS("p { margin: -10px !important; }");

    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that CSSSourceSyntaxHighlighter detects the tokens.
</p>

</body>
</html>
