LayoutTests/inspector/syntax-highlight-css-expected.txt

Tests that CSSSourceSyntaxHighlighter detects the tokens.

a[href='/']: webkit-css-selector
#content > a:hover: *,webkit-css-selector,*,webkit-css-selector,*,webkit-css-selector
@import url(style.css);: webkit-css-at-rule,*,webkit-css-keyword,*
@import url("style.css") projection, tv;: webkit-css-at-rule,*,webkit-css-keyword,*,webkit-css-string,*,webkit-css-keyword,*,webkit-css-keyword,*
@import "/css/fireball_unicode.css"; html {}: webkit-css-at-rule,*,webkit-css-string,*,webkit-css-selector,*
@media screen { body { color: red; } }: webkit-css-at-rule,*,webkit-css-keyword,*,webkit-css-selector,*,webkit-css-property,*,webkit-css-color,*
@font-face { font-family: "MyHelvetica"; }: webkit-css-at-rule,*,webkit-css-property,*,webkit-css-string,*
p { color: color; red: red; color: #000; color: #FFF; color: #123AbC; color: #faebfe; color:papayawhip; }: webkit-css-selector,*,webkit-css-property,*,webkit-css-color,*,webkit-css-property,*,webkit-css-color,*,webkit-css-property,*,webkit-css-color,*,webkit-css-property,*,webkit-css-color,*,webkit-css-property,*,webkit-css-color,*,webkit-css-property,*,webkit-css-color,*
p { margin: -10px !important; }: webkit-css-selector,*,webkit-css-property,*,webkit-css-number,*,webkit-css-important,*
